quiz-backend-api/pyproject.toml
[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
# Specify the test paths
testpaths = ["tests"]

# Add any additional options you want to configure
addopts = "-v --tb=short"

# Specify markers for your tests
markers = [
    "slow: mark test as slow"
]

# Configure other pytest options as needed
log_cli_level = "INFO"  # Set the logging level for the command line

[project]
name = "quiz-backend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "build==1.2.2.post1",
    "cachecontrol==0.14.1",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "cleo==2.1.0",
    "crashtest==0.4.1",
    "distlib==0.3.9",
    "dulwich==0.21.7",
    "fastjsonschema==2.21.1",
    "filelock==3.16.1",
    "idna==3.10",
    "iniconfig==2.0.0",
    "installer==0.7.0",
    "jaraco-classes==3.4.0",
    "keyring==24.3.1",
    "mimesis==18.0.0",
    "more-itertools==10.5.0",
    "msgpack==1.1.0",
    "packaging==24.2",
    "pexpect==4.9.0",
    "pkginfo==1.12.0",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "poetry-core==1.9.1",
    "poetry==1.8.5",
    "poetry-plugin-export==1.8.0",
    "ptyprocess==0.7.0",
    "pycparser==2.22",
    "pyproject-hooks==1.2.0",
    "pytest==8.3.4",
    "rapidfuzz==3.11.0",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "shellingham==1.5.4",
    "tomlkit==0.13.2",
    "trove-classifiers==2024.10.21.16",
    "urllib3==2.3.0",
    "uv==0.5.2",
    "virtualenv==20.28.0",
    "xattr==1.1.0",
    "click==8.2.0",
    "h11==0.16.0",
    "uvicorn==0.34.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "fastapi==0.115.12",
    "pip==25.1.1",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "sqlalchemy==2.0.41",
    "asyncpg==0.30.0",
    "loguru==0.7.3",
    "python-dotenv==1.1.0",
    "deprecated==1.2.18",
    "limits==5.2.0",
    "slowapi==0.1.9",
    "wrapt==1.17.2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "ecdsa==0.19.1",
    "pyasn1==0.6.1",
    "python-jose==3.5.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "alembic==1.16.1",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "greenlet==3.2.3",
    "apscheduler==3.11.0",
    "tzlocal==5.3.1",
    "aiokafka==0.12.0",
    "async-timeout==5.0.1",
    "redis==6.2.0",
    "trio>=0.30.0",
    "pytest-docker>=3.2.3",
    "pytest-mock>=3.14.1",
    "aiosqlite>=0.21.0",
    "pytest-asyncio>=1.1.0",
]
